fix: Move krist related structures to root of crate
This was done because these modeled are not only used for the WebSocket server implementation.
