+++ src/models/webserver/mod.rs
@@ -1,0 +1,2 @@
+pub mod lookup;
+pub mod search;
