--- src/websockets/models/error.rs
@@ -1,8 +1,0 @@
-use serde::{Deserialize, Serialize};
-
-#[derive(Debug, Clone, PartialEq, PartialOrd, Serialize, Deserialize)]
-pub struct ErrorResponse {
-    //pub ok: bool,
-    pub error: String,
-    pub message: Option<String>,
-}
