--- src/websockets/models/webserver/mod.rs
@@ -1,2 +1,0 @@
-pub mod lookup;
-pub mod search;
